On selection, provide the current hierarchy in addition to the selected node
